test: integration

+++ src/interfaces/IUniswapV3PoolVariant.sol
@@ -11,7 +11,7 @@ interface IUniswapV3PoolVariant {
             uint16 observationIndex,
             uint16 observationCardinality,
             uint16 observationCardinalityNext,
-            uint8 feeProtocol,
+            uint32 feeProtocol,
             bool unlocked
         );
